pulled from upstream. and made updates to the campaign details page.
