+++ svgs/CalenderIcon.tsx
@@ -9,7 +9,7 @@ const CalenderIcon = () => {
     >
       <path
         fillRule="evenodd"
-        clip-rule="evenodd"
+        clipRule="evenodd"
         d="M8.16667 4.68799C7.53667 4.71483 7.084 4.77433 6.68383 4.90966C5.99945 5.13881 5.37754 5.52358 4.867 6.03372C4.35646 6.54385 3.9712 7.16546 3.7415 7.84966C3.5 8.57416 3.5 9.46783 3.5 11.2575C3.5 11.3683 3.5 11.4243 3.51517 11.4687C3.52949 11.5116 3.5536 11.5506 3.58559 11.5826C3.61759 11.6146 3.65658 11.6387 3.6995 11.653C3.74383 11.6682 3.79983 11.6682 3.91183 11.6682H24.0882C24.2002 11.6682 24.2562 11.6682 24.3005 11.653C24.3434 11.6387 24.3824 11.6146 24.4144 11.5826C24.4464 11.5506 24.4705 11.5116 24.4848 11.4687C24.5 11.4232 24.5 11.3672 24.5 11.2552C24.5 9.46666 24.5 8.57183 24.2585 7.85083C24.0291 7.16619 23.644 6.54412 23.1334 6.03356C22.6229 5.523 22.0008 5.13788 21.3162 4.90849C20.916 4.77433 20.4622 4.71483 19.8333 4.68799V7.58366C19.8333 8.04779 19.649 8.49291 19.3208 8.8211C18.9926 9.14928 18.5475 9.33366 18.0833 9.33366C17.6192 9.33366 17.1741 9.14928 16.8459 8.8211C16.5177 8.49291 16.3333 8.04779 16.3333 7.58366V4.66699H11.6667V7.58366C11.6667 8.04779 11.4823 8.49291 11.1541 8.8211C10.8259 9.14928 10.3808 9.33366 9.91667 9.33366C9.45254 9.33366 9.00742 9.14928 8.67923 8.8211C8.35104 8.49291 8.16667 8.04779 8.16667 7.58366V4.68799Z"
         fill="#55534E"
       />
