+++ svgs/SearchIcon.tsx
@@ -10,15 +10,15 @@ const SearchIcon = () => {
       <path
         d="M8.375 14.25C11.8959 14.25 14.75 11.3959 14.75 7.875C14.75 4.35413 11.8959 1.5 8.375 1.5C4.85413 1.5 2 4.35413 2 7.875C2 11.3959 4.85413 14.25 8.375 14.25Z"
         stroke="currentColor"
-        stroke-width="1.5"
-        stroke-linejoin="round"
+        strokeWidth="1.5"
+        strokeLinejoin="round"
       />
       <path
         d="M10.4964 5.37863C10.218 5.09972 9.88733 4.87853 9.52329 4.72775C9.15924 4.57698 8.76901 4.49958 8.37498 4.5C7.98094 4.49958 7.59071 4.57698 7.22667 4.72775C6.86262 4.87853 6.53194 5.09972 6.2536 5.37863M12.9582 12.4583L16.1401 15.6401"
         stroke="currentColor"
-        stroke-width="1.5"
-        stroke-linecap="round"
-        stroke-linejoin="round"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
       />
     </svg>
   );
